database/password_generator/requirements.txt
astroid==2.9.3
greenlet==1.1.2
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
msgpack==1.0.3
mysql-connector-python==8.0.28
neovim==0.3.1
platformdirs==2.5.1
protobuf==3.19.4
pylint==2.12.2
pynvim==0.4.3
toml==0.10.2
typing-extensions==4.1.1
wrapt==1.13.3
